requirements.txt
opencv-python==4.9.0.80
playwright==1.41.0
numpy==1.26.3
Pillow==10.2.0
scipy==1.12.0
matplotlib==3.8.2